requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
pyfiglet==0.8.post1
PyYAML==5.4.1
requests==2.26.0
tabulate==0.8.9
urllib3==1.26.6